requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
kombu==5.2.4
pika==1.3.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytube==12.1.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
